refactor: board下の名前空間を整理
