+++ src/main/scala/board/jtag/types/VirtualInstruction.scala
